refactor: move StrokePainter into the base/painter.ts
